nomad/views.py
from django.contrib.auth.models import User
from .models import Cafe
from rest_framework import viewsets
from nomad.serializers import UserSerializer, CafeSerializer
# from django.shortcuts import render


class UserViewSet(viewsets.ModelViewSet):
    """
    API endpoint that allows users to be viewed or edited.
    """
    queryset = User.objects.all().order_by('-date_joined')
    serializer_class = UserSerializer

    
class CafeViewSet(viewsets.ModelViewSet):
    queryset = Cafe.objects.all()
    serializer_class = CafeSerializer
