requirements.txt
Django
psycopg2
gunicorn
djangorestframework
djongo
dnspython
